pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = ""
authors = [
    {name = "nagendrasisodiya",email = "3n3singh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<=3.12"
dependencies = [
    "load-dotenv (>=0.1.0,<0.2.0)",
    "ibm-watsonx-ai (>=1.5.4,<2.0.0)",
    "langchain (>=1.2.13,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-ibm (>=1.0.4,<2.0.0)",
    "langchain-core (>=1.2.20,<2.0.0)",
    "langgraph (>=1.1.3,<2.0.0)",
    "wikipedia (>=1.4.0,<2.0.0)",
    "pandas (<2.4.0)",
    "numpy (>=2.4.3,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.8.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
